requirements-dev.txt
# Development and linting (install with: pip install -r requirements-dev.txt)
pre-commit>=3.6.0
ruff>=0.6.0
pytest-asyncio>=0.23.5
pandas>=2.2.0
